renamed main packge to `at.favre.lib.hkdf` to avoid split packages (see #9)
