+++ src/main/java9/module-info.java
@@ -1,3 +1,3 @@
 module at.favre.lib.hkdf {
-    exports at.favre.lib.crypto; // FIXME split package with other libs, such as at.favre.lib:bcrypt
+    exports at.favre.lib.hkdf;
 }
