+++ src/test/java/at/favre/lib/hkdf/HKDFBenchmarkTest.java
@@ -1,4 +1,4 @@
-package at.favre.lib.crypto;
+package at.favre.lib.hkdf;
 
 import org.apache.commons.lang3.RandomUtils;
 import org.junit.Ignore;
